major refactoring of code and code removal
